src/test/kotlin/Day25Test.kt
import org.eruanno.Day25
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.CsvSource

class Day25Test {
    private val subject = Day25()

    @ParameterizedTest
    @CsvSource(
        "'day-25-test', 3",
        "'day-25', 3249"
    )
    fun testPartOne(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partOne(testFile))
    }

    @ParameterizedTest
    @CsvSource(
        "'day-25-test', '0'",
        "'day-25', '0'"
    )
    fun testPartTwo(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partTwo(testFile))
    }
}
